requirements.txt
torch==2.3.1
torchvision==0.18.1
numpy==1.26.4
Pillow==10.3.0
pyyaml==6.0.1
tqdm==4.66.4
scikit-learn==1.5.0
